Fix: #109-호흡 화면 영상 자막 싱크 맞추기

+++ Spha/Shared/Managers/BreathingManager.swift
@@ -12,12 +12,12 @@ protocol BreathingManager: ObservableObject {
     var isBreathingCompleted: Bool { get set }
     var phaseText: String { get set } 
     var showText: Bool { get set }
-    var timerCount: Int { get set }
+    var timerCount: Double { get set }
     var showTimer: Bool { get set }
     var activeCircle: Int { get set }
 
     func startBreathingIntro() // 초기 인트로 시작
     func startBreathingCycle()// 호흡 사이클 시작
-    func startPhase(phase: BreathingPhase, duration: Int) async // 특정 호흡 단계 시작
+    func startPhase(phase: BreathingPhase, duration: Double) async // 특정 호흡 단계 시작
 }
